[Backend] Separated all middlewares in the index file of the API to separate files and functions
