+++ Backend/src/middlewares/index.ts
@@ -1,0 +1,3 @@
+export {default as getRouteDetailsMiddleware} from "./route-information";
+export {default as assureRouteBasicInfo} from "./route-basic-info"
+export {default as assureBodyFields} from "./body-fields";
